+++ falsy_async_guard.d.ts
@@ -1,0 +1,2 @@
+export declare function falsy_async_guard<O extends unknown = unknown>(label: string, ...fns: Function[]): Promise<O>;
+export { falsy_async_guard as falsish_async_guard };