+++ falsy_guard.d.ts
@@ -1,0 +1,2 @@
+export declare function falsy_guard<O extends unknown = unknown>(label: string, ...fns: Function[]): O;
+export { falsy_guard as falsish_guard };