+++ nullish.ts
@@ -1,0 +1,2 @@
+export type nullish = null|undefined
+export type nullsy = nullish