+++ nullish_async_guard.d.ts
@@ -1,0 +1,2 @@
+export declare function nullish_async_guard<O extends unknown = unknown>(label: string, ...fns: Function[]): Promise<O>;
+export { nullish_async_guard as nullsy_async_guard };