+++ nullish_guard.d.ts
@@ -1,0 +1,2 @@
+export declare function nullish_guard<O extends unknown = unknown>(label: string, ...fns: Function[]): O;
+export { nullish_guard as nullsy_guard };