+++ nullsy.ts
@@ -1,1 +1,0 @@
-export type nullsy = null|undefined